+++ packages/core/src/compiler/StyleFunctionCaller.ts
@@ -1,24 +1,19 @@
-import * as ad from "../../types/ad";
-import { SourceRange } from "../../types/ast";
-import {
-  CompFunc,
-  ConstrFunc,
-  FuncParam,
-  ObjFunc,
-} from "../../types/functions";
+import * as ad from "../types/ad";
+import { SourceRange } from "../types/ast";
+import { CompFunc, ConstrFunc, FuncParam, ObjFunc } from "../types/functions";
 
 import { Result } from "true-myth";
-import { Context } from "../../shapes/Samplers";
-import { Shape } from "../../shapes/Shapes";
-import { StyleError } from "../../types/errors";
-import { ArgValWithSourceLoc, Value } from "../../types/value";
+import { Context } from "../shapes/Samplers";
+import { Shape } from "../shapes/Shapes";
+import { StyleError } from "../types/errors";
+import { ArgValWithSourceLoc, Value } from "../types/value";
 import {
   badArgumentTypeError,
   functionInternalError,
   missingArgumentError,
   tooManyArgumentsError,
-} from "../../utils/Error";
-import { checkType } from "../styleTypeChecker/styleTypeChecker";
+} from "../utils/Error";
+import { checkType } from "./StyleTypeChecker";
 
 const { ok, err } = Result;
